fix: check delete before delete

+++ .erb/scripts/clean.js
@@ -1,4 +1,5 @@
 import rimraf from 'rimraf';
+import fs from 'fs';
 import webpackPaths from '../configs/webpack.paths';
 
 const foldersToRemove = [
@@ -8,5 +9,5 @@ const foldersToRemove = [
 ];
 
 foldersToRemove.forEach((folder) => {
-  rimraf.sync(folder);
+  if (fs.existsSync(folder)) rimraf.sync(folder);
 });
